Update for Webpack 4.
This basically re-writes the plugin based on the BannerPlugin built
in to Webpack 4. However, it does remove the unneeded (and
problematic) text substitution in BannerPlugin and preserves the
options syntax of the current version.

Also now supported are "include"/"exclude" options that can be used to
limit the prepending to only certain chunk files.